+++ feline/Element.h
@@ -12,25 +12,25 @@ class Element
 	Matrix3d stiffnessMat;
 	Matrix3d undeformShapeMat,
 		     undeformShapeMatInv;
-	GenMatrix<float,6,12> strainMat;
-	GenMatrix<float,6,6> matConstantsMat;
-	GenMatrix<float,12,12> undeformStiffnessMat,
+	GenMatrix<double,6,12> strainMat;
+	GenMatrix<double,6,6> matConstantsMat;
+	GenMatrix<double,12,12> undeformStiffnessMat,
 						   massMat;
-	float undeformVolume;
-	float E, v; // E is Young's Modulus, v is Poisson's Ratio
-	float density;
+	double undeformVolume;
+	double E, v; // E is Young's Modulus, v is Poisson's Ratio
+	double density;
 	void preCompute();
 	void preComputeUndeformedStiffnessMat();
 	void preComputeMassMat();
 
 public:
-	Element(Node* n1, Node* n2, Node* n3, Node* n4, float _E, float _v, float _density);
+	Element(Node* n1, Node* n2, Node* n3, Node* n4, double _E, double _v, double _density);
 	Matrix3d computeDeformationMat();
 	//computeundeformstiffnessmat B^T * c * B where B = LN
 	//computestiffnessmat dfe/dx |x = x*
 
-	GenMatrix<float,12,12>* getStiffnessMat(){return &undeformStiffnessMat;}
-	GenMatrix<float,12,12>* getMassMat(){return &massMat;}
+	GenMatrix<double,12,12>* getStiffnessMat(){return &undeformStiffnessMat;}
+	GenMatrix<double,12,12>* getMassMat(){return &massMat;}
 	void renderElement();
 
 	~Element(void);
